requirements.txt
deepspeed==0.10.0
transformers==4.30.2
torch==2.0.1
datasets==2.13.0
wandb==0.15.4
optuna==3.2.0
nltk==3.8.1
kubernetes==26.1.0
pycryptodome==3.18.0
scikit-learn==1.2.2
python-dotenv==1.0.0